+++ RTS Project code/Motor2D/Slide.cpp
@@ -102,22 +102,22 @@ void Slide::HorizontalSlide()
 	{
 		if (!non_lerp)
 		{
-			screen.x = Lerp(-screen.w, 0, current_cutoff);									// Horizontal linearly interpolated slide coming from the left.
+			screen.x = (int)Lerp((float)-screen.w, 0.f, current_cutoff);									// Horizontal linearly interpolated slide coming from the left.
 		}
 		else
 		{
-			screen.x = N_Lerp(-screen.w, 0, current_cutoff);								// Horizontal non-linearly interpolated slide coming from the left.
+			screen.x = (int)N_Lerp((float)-screen.w, 0.f, current_cutoff);								// Horizontal non-linearly interpolated slide coming from the left.
 		}
 	}
 	else
 	{
 		if (!non_lerp)
 		{
-			screen.x = Lerp(screen.w, 0, current_cutoff);									// Horizontal linearly interpolated slide coming from the right.
+			screen.x = (int)Lerp((float)screen.w, 0.f, current_cutoff);									// Horizontal linearly interpolated slide coming from the right.
 		}
 		else
 		{
-			screen.x = N_Lerp(screen.w, 0, current_cutoff);									// Horizontal non-linearly interpolated slide coming from the right.
+			screen.x = (int)N_Lerp((float)screen.w, 0.f, current_cutoff);									// Horizontal non-linearly interpolated slide coming from the right.
 		}
 	}
 }
@@ -128,29 +128,29 @@ void Slide::VerticalSlide()
 	{
 		if (!non_lerp)
 		{
-			screen.y = Lerp(-screen.h, 0, current_cutoff);									// Vertical linearly interpolated slide coming from the top.
+			screen.y = (int)Lerp((float)-screen.h, 0.f, current_cutoff);									// Vertical linearly interpolated slide coming from the top.
 		}
 		else
 		{
-			screen.y = N_Lerp(-screen.h, 0, current_cutoff);								// Vertical non-linearly interpolated slide coming from the top.
+			screen.y = (int)N_Lerp((float)-screen.h, 0.f, current_cutoff);								// Vertical non-linearly interpolated slide coming from the top.
 		}
 	}
 	else
 	{
 		if (!non_lerp)
 		{
-			screen.y = Lerp(screen.h, 0, current_cutoff);									// Vertical linearly interpolated slide coming from the bottom.
+			screen.y = (int)Lerp((float)screen.h, 0.f, current_cutoff);									// Vertical linearly interpolated slide coming from the bottom.
 		}
 		else
 		{
-			screen.y = N_Lerp(screen.h, 0, current_cutoff);									// Vertical non-linearly interpolated slide coming from the bottom.
+			screen.y = (int)N_Lerp((float)screen.h, 0.f, current_cutoff);									// Vertical non-linearly interpolated slide coming from the bottom.
 		}
 	}
 }
 
 void Slide::DrawSlide()
 {
-	SDL_SetRenderDrawColor(App->render->renderer, slide_colour.r, slide_colour.g, slide_colour.b, 255);
+	SDL_SetRenderDrawColor(App->render->renderer, (Uint8)slide_colour.r, (Uint8)slide_colour.g, (Uint8)slide_colour.b, (Uint8)255);
 	SDL_RenderFillRect(App->render->renderer, &screen);
 }
