+++ RTS Project code/Motor2D/Tree.cpp
@@ -74,7 +74,7 @@ bool Tree::CleanUp()
 
 void Tree::Draw()
 {
-	App->render->Blit(entity_sprite, pixel_position.x, pixel_position.y - 20, blit_section);
+	App->render->Blit(entity_sprite, (int)pixel_position.x, (int)pixel_position.y - 20, blit_section);
 }
 
 void Tree::InitEntity()
@@ -85,8 +85,8 @@ void Tree::InitEntity()
 
 	iPoint world_position = App->map->MapToWorld(tile_position.x, tile_position.y);
 
-	pixel_position.x = world_position.x;
-	pixel_position.y = world_position.y;
+	pixel_position.x = (float)world_position.x;
+	pixel_position.y = (float)world_position.y;
 
 	tiles_occupied_x = 1;
 	tiles_occupied_y = 1;
